requirements.txt
MySQL-python==1.2.5
Django==1.11.2
djoser==0.6.0
djet==0.2.1
coverage==4.3.1
djangorestframework-jwt
django-password-validation==0.1.1
django-cors-headers
djangorestframework
dj-database-url
dj-static
django-appconf
django-compressor
drf-nested-routers
gunicorn
six
static3
wsgiref
